requirements.txt
google-ads==26.0.0
requests
pytz
